+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2024-08-01
+
+- add debug command to troubleshoot future errors
+- add local timestamp to events for easy reference
+- fix log level processing
+- fix protocol error when sending data to webhook
+- update content for easier copy on mobile
+- update default log level to error
+- update default status messages
+- update references for 7dtd-1.0-b333
+- update status events to be more compact
+
 ## [1.0.2] - 2023-11-21
 
 - add setup section to readme